airguard/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openmeteo-requests==1.3.0
requests-cache==1.2.0
retry-requests==2.0.0
pandas==2.2.2
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.2
scipy==1.13.0
pykrige==1.7.1
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
